requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
executing==2.0.1
idna==3.6
inflection==0.5.1
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg-binary==3.1.13
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==7.7.0
traitlets==5.14.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
wcwidth==0.2.12

celery~=5.3.6